add swagger documentation

+++ app/Http/Controllers/UserController.php
@@ -5,21 +5,36 @@ namespace App\Http\Controllers;
 use Illuminate\Http\Request;
 use App\Http\Controllers\BaseController as BaseController;
 use App\Models\User;
+use Ramsey\Uuid\Type\Integer;
 use Validator;
 use App\Http\Resources\UserResource;
 
 class UserController extends BaseController
 {
     /**
+     * @OA\Info(
+     *      version="1.0.0",
+     *      title="Laravel API",
+     *      description="",
+     *      @OA\Contact(
+     *          email="karolpyt1999@gmail.com"
+     *      ),
+     * )
+     *
      * Display a listing of the resource.
      *
      * @return \Illuminate\Http\Response
+     * @OA\Get(
+     *     path="/users",
+     *     description="Gets a list of users",
+     *     @OA\Response(response="default", description="Welcome page")
+     * )
      */
+
+
     public function index()
     {
-
         $products = User::all();
-
         return $this->sendResponse(UserResource::collection($products), 'Users retrieved successfully.');
     }
 
@@ -32,11 +47,9 @@ class UserController extends BaseController
     public function show($id)
     {
         $user = User::find($id);
-
         if (is_null($user)) {
             return $this->sendError('User not found.');
         }
-
         return $this->sendResponse(new UserResource($user), 'User retrieved successfully.');
     }
 
@@ -47,9 +60,13 @@ class UserController extends BaseController
      * @param  int  $id
      * @return \Illuminate\Http\Response
      */
-    public function update(Request $request, User $user)
+    public function update(Request $request, int $id)
     {
         $input = $request->all();
+        $user = User::find($id);
+        if (is_null($user)) {
+            return $this->sendError('User not found.');
+        }
         $validator = Validator::make($input, [
             'email' => 'email'
         ]);
@@ -59,7 +76,6 @@ class UserController extends BaseController
         }
 
         $user->update(array_filter($request->all()));
-
         return $this->sendResponse(new UserResource($user), 'User updated successfully.');
     }
 
@@ -69,10 +85,13 @@ class UserController extends BaseController
      * @param  int  $id
      * @return \Illuminate\Http\Response
      */
-    public function destroy(User $user)
+    public function destroy($id)
     {
+        $user = User::find($id);
+        if (is_null($user)) {
+            return $this->sendError('User not found.');
+        }
         $user->delete();
-
         return $this->sendResponse([], 'User deleted successfully.');
     }
 }
